Added Flask code
This is the code for the Flask app running all the main website code. I originally had this as part of running the website on PythonAnywhere, and the JavaScript code still pulls JSON values from this app when it's running. If you're not going to use PythonAnywhere, try to figure out how to integrate this with the front-end code, or find a way to have the front-end code run without Flask.